homebudget-source/homebudget/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openpyxl==3.1.5
pandas==2.2.3
python-multipart==0.0.12
anthropic>=0.40.0
pyyaml==6.0.2
watchdog==5.0.2
requests==2.32.3
xlsxwriter==3.2.0
